add new user table component to the users component
