+++ src/app/features/workplace/users/users.component.html
@@ -29,49 +29,12 @@
   </div>
 </div>
 <ng-container *ngIf="users.length > 0; else noUsers">
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
-      <table class="govuk-table">
-        <thead class="govuk-table__head">
-          <tr class="govuk-table__row">
-            <th class="govuk-table__header" scope="col">Full name</th>
-            <th class="govuk-table__header" scope="col">Username</th>
-            <th class="govuk-table__header" scope="col">Last updated</th>
-            <th class="govuk-table__header" scope="col">Permissions</th>
-            <th class="govuk-table__header" scope="col">Status</th>
-          </tr>
-        </thead>
-        <tbody class="govuk-table__body">
-          <ng-container *ngFor="let user of users">
-            <tr class="govuk-table__row" [class.govuk-error-table__row--highlight]="isPending(user)">
-              <td class="govuk-table__cell" [ngClass]="{ 'govuk-!-padding-left-3': isPending(user) }">
-                <ng-container *ngIf="canViewUser; else noLink">
-                  <a [routerLink]="['/workplace', workplace.uid, 'user', user.uid]" data-testid="username-link">{{
-                    user.fullname
-                  }}</a>
-                </ng-container>
-                <ng-template #noLink>
-                  {{ user.fullname }}
-                </ng-template>
-              </td>
-              <td class="govuk-table__cell">
-                {{ isPending(user) ? '-' : user.username }}
-              </td>
-              <td class="govuk-table__cell">
-                {{ user.updated | date: 'd MMM y' }}
-              </td>
-              <td class="govuk-table__cell">
-                {{ getUserType(user) }}
-              </td>
-              <td class="govuk-table__cell">
-                {{ user.status }}
-              </td>
-            </tr>
-          </ng-container>
-        </tbody>
-      </table>
-    </div>
-  </div>
+  <app-user-table
+    [users]="users"
+    [canViewUser]="canViewUser"
+    [workplace]="workplace"
+    [userPermissionsTypes]="userPermissionsTypes"
+  ></app-user-table>
 </ng-container>
 
 <ng-template #noUsers>
