+++ src/app/shared/components/users-table/user-table.component.spec.ts
@@ -1,9 +1,11 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { RouterModule } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
+import { Roles } from '@core/model/roles.enum';
 import { UserDetails } from '@core/model/userDetails.model';
 import { AdminManagerUser, AdminUser, PendingAdminUser } from '@core/test-utils/admin/MockAdminService';
 import { EditUser, ReadUser } from '@core/test-utils/MockUserService';
+import { getUserPermissionsTypes } from '@core/utils/users-util';
 import { SharedModule } from '@shared/shared.module';
 import { render } from '@testing-library/angular';
 
@@ -13,20 +15,22 @@ import { UserTableComponent } from './user.table.component';
 describe('UserTableComponent', () => {
   const userArr = [ReadUser(), EditUser(), EditUser()] as UserDetails[];
   const adminUserArr = [AdminUser(), AdminManagerUser(), PendingAdminUser()] as UserDetails[];
+  const permissionTypes = getUserPermissionsTypes(true);
 
   const setup = async (admin = false, canViewUser = true) => {
-    const { fixture, getByText, getByTestId } = await render(UserTableComponent, {
+    const { fixture, getByText, getByTestId, queryByText, queryAllByText } = await render(UserTableComponent, {
       imports: [SharedModule, RouterModule, RouterTestingModule, HttpClientTestingModule],
       componentProperties: {
         workplace: !admin && Establishment,
         users: admin ? adminUserArr : userArr,
         canViewUser: canViewUser,
+        userPermissionsTypes: !admin && permissionTypes,
       },
       declarations: [],
     });
     const component = fixture.componentInstance;
 
-    return { component, fixture, getByText, getByTestId };
+    return { component, fixture, getByText, getByTestId, queryByText, queryAllByText };
   };
 
   it('should render a User Account Summary Workplace Component', async () => {
@@ -87,6 +91,99 @@ describe('UserTableComponent', () => {
     adminUsers.forEach((adminUser) => {
       const user = getByText(adminUser.fullname);
       expect(user.getAttribute('href')).toEqual(`/sfcadmin/users/${adminUser.uid}`);
+    });
+  });
+
+  describe('Permissions column', () => {
+    it('should have permission as Admin or AdminManagers if the users are an admin or an admin manager', async () => {
+      const { queryAllByText, queryByText } = await setup(true);
+
+      expect(queryAllByText('Admin')).toBeTruthy();
+      expect(queryByText('AdminManager')).toBeTruthy();
+    });
+
+    it('should have permission as Primary edit and WDF when user isPrimary and canManageWdfClaims are true and role is Edit', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'Edit' as Roles;
+      component.users[0].isPrimary = true;
+      component.users[0].canManageWdfClaims = true;
+
+      fixture.detectChanges();
+
+      expect(queryByText('Primary edit and WDF')).toBeTruthy();
+    });
+
+    it('should have permission as Primary edit when user isPrimary is true, canManageWdfClaims is false and role is Edit', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'Edit' as Roles;
+      component.users[0].isPrimary = true;
+      component.users[0].canManageWdfClaims = false;
+
+      fixture.detectChanges();
+
+      expect(queryByText('Primary edit')).toBeTruthy();
+    });
+
+    it('should have permission as Edit and WDF when user isPrimary is false, canManageWdfClaims is true and role is Edit', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'Edit' as Roles;
+      component.users[0].isPrimary = false;
+      component.users[0].canManageWdfClaims = true;
+
+      fixture.detectChanges();
+
+      expect(queryByText('Edit and WDF')).toBeTruthy();
+    });
+
+    it('should have permission as Edit when user isPrimary is false, canManageWdfClaims is false and role is Edit', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'Edit' as Roles;
+      component.users[0].isPrimary = false;
+      component.users[0].canManageWdfClaims = false;
+
+      fixture.detectChanges();
+
+      expect(queryByText('Edit')).toBeTruthy();
+    });
+
+    it('should have permission as Read only and WDF when user canManageWdfClaims is true and role is Read', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'Read' as Roles;
+      component.users[0].canManageWdfClaims = true;
+      component.users[0].isPrimary = false;
+
+      fixture.detectChanges();
+
+      expect(queryByText('Read only and WDF')).toBeTruthy();
+    });
+
+    it('should have permission as Read only when user canManageWdfClaims is false and role is Read', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'Read' as Roles;
+      component.users[0].canManageWdfClaims = false;
+      component.users[0].isPrimary = false;
+
+      fixture.detectChanges();
+
+      expect(queryByText('Read only')).toBeTruthy();
+    });
+
+    it('should have permission as WDF when user canManageWdfClaims is true and role is None', async () => {
+      const { component, fixture, queryByText } = await setup();
+
+      component.users[0].role = 'None' as Roles;
+      component.users[0].canManageWdfClaims = true;
+      component.users[0].isPrimary = false;
+
+      fixture.detectChanges();
+
+      expect(queryByText('WDF')).toBeTruthy();
     });
   });
 });
