+++ src/app/shared/components/users-table/user.table.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { Establishment } from '@core/model/establishment.model';
-import { UserDetails, UserStatus } from '@core/model/userDetails.model';
+import { UserDetails, UserPermissionsType, UserStatus } from '@core/model/userDetails.model';
 
 @Component({
   selector: 'app-user-table',
@@ -10,8 +10,23 @@ export class UserTableComponent {
   @Input() users: UserDetails[] = [];
   @Input() canViewUser = true;
   @Input() workplace: Establishment;
+  @Input() userPermissionsTypes: UserPermissionsType[];
 
   public isPending(user: UserDetails): boolean {
     return user.status === UserStatus.Pending;
+  }
+
+  public getUserType(user: UserDetails): string {
+    if (this.workplace) {
+      const userType = this.userPermissionsTypes.find(
+        (type) =>
+          type.role === user.role &&
+          type.canManageWdfClaims === user.canManageWdfClaims &&
+          !!user.isPrimary === !!type.isPrimary,
+      );
+
+      return userType?.userTableValue;
+    }
+    return user.role;
   }
 }
